move examination page stores
