+++ store/examine/mock.ts
