+++ store/examine/options.ts
